requirements.txt
numpy
pandas
mathematics
sklearn
quandl
matplotlib

netCDF4

Cartopy

request
python-oauth2
